Updated count from 32 to 64 bit number
git-svn-id: https://satscm.esac.esa.int/esasky/esasky-app/trunk@5227 aa7853a0-da02-4334-bf23-07ffea9f4f01

+++ esasky-cl/src/main/java/esac/archive/esasky/cl/web/client/api/Api.java
@@ -650,7 +650,7 @@ public class Api {
 		}else {
 			countStatus.registerObserver(new CountObserver() {
 				@Override
-				public void onCountUpdate(int newCount) {
+				public void onCountUpdate(long newCount) {
 					JSONObject obsCount = new  JSONObject();
 					
 					for(ExtTapDescriptor desc : DescriptorRepository.getInstance().getExtTapDescriptors().getDescriptors()) {
@@ -686,7 +686,7 @@ public class Api {
 		}else {
 			countStatus.registerObserver(new CountObserver() {
 				@Override
-				public void onCountUpdate(int newCount) {
+				public void onCountUpdate(long newCount) {
 					JSONObject obsCount = new  JSONObject();
 					
 					for (BaseDescriptor currObs : descriptors.getDescriptors()) {
