+++ esasky-cl/src/main/java/esac/archive/esasky/cl/web/client/status/CountObserver.java
@@ -1,5 +1,5 @@
 package esac.archive.esasky.cl.web.client.status;
 
 public interface CountObserver {
-	void onCountUpdate(int newCount);
+	void onCountUpdate(long newCount);
 }
